uprobe_helpers: Fix path acquisition function
Get the file path from `readlink /proc/*/exe`, which may follow at the end with
`(deleted)`,

$ sudo readlink /proc/1135/exe
/usr/bin/python 3.7 (deleted)

When the file has been deleted, file path cannot be obtained correctly.
This is common in operating systems, such as updating a software package
with `rpm -Uvh python.rpm`, which can cause this problem. The solution is to
simply ignore the following `(deleted)`.

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I44c3a1c7c8a65cfba4d12d13983f83171805082a
Reviewed-on: http://gerrit.kylin.com/c/ketones/+/63829