libbpf-tools: Allow tcppktlat to run on old kernels
Currently, tcppktlat can only run on kernel v5.12+
because of the usages of bpf_get_socket_cookie().
This commit gets the tool run on old kernels in a CO-RE
way. While at it, also adds tcp_destroy_sock tracepoint.
This is necessary since there are cases where the socket
does not get the chance to reach the tcp_rcv_space_adjust
tracepoint and entries in start map never get removed.

Signed-off-by: Hengqi Chen <hengqi.chen@gmail.com>
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: Iad71643285a848b107310e63b0dbc76038abac84
Reviewed-on: http://gerrit.kylin.com/c/ketones/+/66010